Fix: move timeline logic to component
